# openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "FastAPI", "version": "0.1.0" },
  "paths": {
...
    "/v1/warehouses": {
      "get": {
        "summary": "List Warehouses",
        "operationId": "list_warehouses_v1_warehouses_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WarehouseList" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Warehouse",
        "operationId": "create_warehouse_v1_warehouses_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_warehouse_v1_warehouses_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Warehouse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/v1/warehouses/{warehouse_id}": {
      "post": {
        "summary": "Update Warehouse",
        "operationId": "update_warehouse_v1_warehouses__warehouse_id__post",
        "parameters": [
          {
            "name": "warehouse_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Warehouse Id" }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Body_update_warehouse_v1_warehouses__warehouse_id__post"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Warehouse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Retrieve Warehouse",
        "operationId": "retrieve_warehouse_v1_warehouses__warehouse_id__get",
        "parameters": [
          {
            "name": "warehouse_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Warehouse Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Warehouse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Warehouse",
        "operationId": "delete_warehouse_v1_warehouses__warehouse_id__delete",
        "parameters": [
          {
            "name": "warehouse_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Warehouse Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WarehouseDelete" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_create_warehouse_v1_warehouses_post": {
        "properties": {
          "name": { "type": "string", "title": "Name" },
          "active": { "type": "string", "title": "Active" },
          "phone": { "type": "string", "title": "Phone" },
          "address_line1": { "type": "string", "title": "Address Line1" },
          "address_line2": { "type": "string", "title": "Address Line2" },
          "address_city": { "type": "string", "title": "Address City" },
          "address_state": { "type": "string", "title": "Address State" },
          "address_country": { "type": "string", "title": "Address Country" },
          "address_postal_code": {
            "type": "string",
            "title": "Address Postal Code"
          }
        },
        "type": "object",
        "required": [
          "name",
          "active",
          "phone",
          "address_line1",
          "address_line2",
          "address_city",
          "address_state",
          "address_country",
          "address_postal_code"
        ],
        "title": "Body_create_warehouse_v1_warehouses_post"
      },
...
      "Body_update_warehouse_v1_warehouses__warehouse_id__post": {
        "properties": {
          "name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Name"
          },
          "active": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Active"
          },
          "phone": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Phone"
          },
          "address_line1": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Address Line1"
          },
          "address_line2": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Address Line2"
          },
          "address_city": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Address City"
          },
          "address_state": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Address State"
          },
          "address_country": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Address Country"
          },
          "address_postal_code": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Address Postal Code"
          }
        },
        "type": "object",
        "title": "Body_update_warehouse_v1_warehouses__warehouse_id__post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
      "Warehouse": {
        "properties": {
          "created": { "type": "integer", "title": "Created" },
          "updated": { "type": "integer", "title": "Updated" },
          "livemode": { "type": "boolean", "title": "Livemode" },
          "name": { "type": "string", "title": "Name" },
          "active": { "type": "boolean", "title": "Active" },
          "phone": { "type": "string", "title": "Phone" },
          "address": { "$ref": "#/components/schemas/WarehouseAddress" },
          "id": { "type": "string", "title": "Id" },
          "object": {
            "type": "string",
            "title": "Object",
            "default": "warehouse"
          }
        },
        "type": "object",
        "required": [
          "created",
          "updated",
          "livemode",
          "name",
          "active",
          "phone",
          "address",
          "id"
        ],
        "title": "Warehouse"
      },
      "WarehouseAddress": {
        "properties": {
          "line1": { "type": "string", "title": "Line1" },
          "line2": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Line2"
          },
          "city": { "type": "string", "title": "City" },
          "state": { "type": "string", "title": "State" },
          "country": { "type": "string", "title": "Country" },
          "postal_code": { "type": "string", "title": "Postal Code" }
        },
        "type": "object",
        "required": ["line1", "city", "state", "country", "postal_code"],
        "title": "WarehouseAddress"
      },
      "WarehouseDelete": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "object": {
            "type": "string",
            "title": "Object",
            "default": "warehouse"
          },
          "deleted": { "type": "boolean", "title": "Deleted" }
        },
        "type": "object",
        "required": ["id", "deleted"],
        "title": "WarehouseDelete"
      },
      "WarehouseList": {
        "properties": {
          "object": { "type": "string", "title": "Object", "default": "list" },
          "url": {
            "type": "string",
            "title": "Url",
            "default": "/v1/warehouses"
          },
          "has_more": { "type": "boolean", "title": "Has More" },
          "data": {
            "items": { "$ref": "#/components/schemas/Warehouse" },
            "type": "array",
            "title": "Data",
            "default": []
          }
        },
        "type": "object",
        "required": ["has_more"],
        "title": "WarehouseList"
      }
    }
  }
}